move plugin system to FreeMote.Psb
