+++ FreeMote.Psb/FreeMote.Psb.csproj
@@ -21,7 +21,7 @@
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <LangVersion>7.3</LangVersion>
+    <LangVersion>8</LangVersion>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
@@ -31,25 +31,40 @@
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
-    <LangVersion>7.3</LangVersion>
+    <LangVersion>8</LangVersion>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <DocumentationFile>bin\Release\FreeMote.Psb.xml</DocumentationFile>
   </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
+    <Reference Include="System.ComponentModel.Composition" />
     <Reference Include="System.Core" />
     <Reference Include="System.Drawing" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AudioMetadata.cs" />
+    <Compile Include="IPsbType.cs" />
+    <Compile Include="IResourceMetadata.cs" />
+    <Compile Include="Plugins\FreeMount.cs" />
+    <Compile Include="Plugins\FreeMountContext.cs" />
+    <Compile Include="Plugins\IPsbAudioFormatter.cs" />
+    <Compile Include="Plugins\IPsbImageFormatter.cs" />
+    <Compile Include="Plugins\IPsbKeyProvider.cs" />
+    <Compile Include="Plugins\IPsbPluginInfo.cs" />
+    <Compile Include="Plugins\IPsbShell.cs" />
     <Compile Include="PSB.cs" />
     <Compile Include="PsbHelper.cs" />
     <Compile Include="PsbPainter.cs" />
     <Compile Include="PsbResCollector.cs" />
     <Compile Include="PsbTypes.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ResourceMetadata.cs" />
+    <Compile Include="ImageMetadata.cs" />
+    <Compile Include="Types\ImageType.cs" />
+    <Compile Include="Types\MmoType.cs" />
+    <Compile Include="Types\MotionType.cs" />
+    <Compile Include="Types\SoundArchiveType.cs" />
     <Compile Include="Textures\TextureCombiner.cs" />
     <Compile Include="Textures\TexturePacker.cs" />
     <Compile Include="Textures\TextureSpliter.cs" />
@@ -79,5 +94,6 @@
       <Link>.editorconfig</Link>
     </None>
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
