+++ FreeMote.Psb/IPsbType.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using FreeMote.Psb.Types;
+
+namespace FreeMote.Psb
+{
+    interface IPsbType
+    {
+        PsbType PsbType { get; }
+
+        bool IsThisType(PSB psb);
+    }
+}
