+++ FreeMote.Psb/Plugins/IPsbAudioFormatter.cs
