+++ FreeMote.Psb/Plugins/IPsbImageFormatter.cs
