+++ FreeMote.Psb/Plugins/IPsbPluginInfo.cs
