+++ FreeMote.Psb/Plugins/IPsbShell.cs
