+++ FreeMote.Psb/Types/SoundArchiveType.cs
@@ -1,0 +1,11 @@
+﻿namespace FreeMote.Psb.Types
+{
+    class SoundArchiveType : IPsbType
+    {
+        public PsbType PsbType => PsbType.SoundArchive;
+        public bool IsThisType(PSB psb)
+        {
+            throw new System.NotImplementedException();
+        }
+    }
+}
